refactor criteria folder
