+++ apps/modernization-ui/src/design-system/date/criteria/date-criteria.module.scss
@@ -1,4 +1,4 @@
-.options-wrapper {
+.operators {
     display: flex;
     gap: 1rem;
 }
