+++ apps/modernization-ui/src/design-system/date/criteria/range/validateDateRange.ts
@@ -1,6 +1,6 @@
 import { Validator } from 'validation';
-import { asDate, asDateEntry, DateBetweenCriteria } from './entry';
-import { validateDate } from './validateDate';
+import { asDate, asDateEntry, validateDate } from 'design-system/date';
+import { DateBetweenCriteria } from 'design-system/date/criteria';
 
 /**
  * Validates that the given "from" and "to" dates are valid and that the "to" date is not earlier than the "from" date.
